refactor: create new ExecutorProcessor class
